Upgrading to django 1.6
Removed the gdata package.  We weren't using it anymore and it generated
a warning.

Removed use of simplejson

Wraped some extra code inside transaction.atomic().  In general we
should do that when we expect an IntegrityError to be raised.  This puts
the code inside a savepoint, rather than the overall transaction.

Removed the compatability monkeypatches, we now have the real versions
from django.