Fix for repositioning sync helpers
Before both the SubtitleRepeatDirective and WorkingSubtitlesController
both hooked up to the same signal.  If a subtitle was inserted/removed,
then sync helpers would get reposition correctly only if the
SubtitleRepeatDirective calback fired first, which wasn't happening for
me.

Changed it so WorkingSubtitlesController hooks up to a signal from
SubtitleRepeatDirective and it only updates the sync helpers after the
subtitle list has been updated.