New way to handle updates for timing adjustments via dragging
Made it so whenever you drag, we update the subtitle list -- this is how
it used to work.

The issue with that is undo.  You don't want to create a new undo entry
for every mousemove event.  Added some code to make it so we merge the
undo entries into a single entry.

pculture/unisubs#3233