Updated the GUITests
Made it so the tests themselves don't access the database.  I was having
a lot of trouble setting that up on our kubernetes stack.  Instead, we
now run the setup_guitests management commands before starting the
webserver.  That command is responsible for initializing the DB before
the tests.  One nice feature of this change is it's easier to step
through tests manually.

- Added ability to run `dev up` and `dev manage` against the GUI test
  environment
- Fixed some issues with disabling the captchas for GUI tests.
- Moved the wait_for_database code to setup_guitests.
- Updated the docs