Removing last bits of celery code
Updated scheduler code to only remove periodic tasks.  This way we can
use enqueue_in() and not worry about it getting deleted.

Added the utils.taskqueue module which wraps rq/rq-scheduler/django-rq.
I think this is nice since it means if we ever want to switch task
queues again, we only need to update the code there, not the entire
codebase.  It also lets us control how things work in the unittests
better.