Updating form template code to make render_field a tag rather
It's better to have it a tag rather than a field because it makes it
possible to take extra arguments.  It's a lot cleaner to pass a tag
reverse_required than to have 2 separate filters.

Also, flatten out the variables we pass to the field.html template.
This makes it more flexible, since you can also pass in other things
than an actual Field objects.